official release publication (wip)

+++ scripts/ciPublish.sh
@@ -10,10 +10,20 @@ VERSION_NUMBER=$(<"$file")
 
 echo "Publish npm for $VERSION_NUMBER"
 
-docker build --rm -f scripts/docker/Dockerfile-publish.build -t cytomine/cytomine-js-client-publish \
-        --build-arg VERSION_NUMBER=$VERSION_NUMBER  \
-        --build-arg NPM_KEY=$NPM_KEY  .
+if [[ $VERSION_NUMBER =~ [0-9]+.[0-9]+.[0-9]$ ]]; then
+  echo "Official release"
 
-containerId=$(docker create cytomine/cytomine-js-client-publish)
+  docker build --rm -f scripts/docker/Dockerfile-publish.build -t cytomine/cytomine-js-client-publish \
+          --build-arg VERSION_NUMBER=$VERSION_NUMBER  \
+          --build-arg NPM_KEY=$NPM_KEY  .
 
-docker rm $containerId
+  containerId=$(docker create cytomine/cytomine-js-client-publish)
+
+  docker rm $containerId
+
+else
+  echo "Snapshot, not publishing it...to be implemented"
+fi
+
+
+
